requirements.txt
numpy
omegaconf==2.3.0
thop==0.1.1-2209072238
Pillow==9.5.0
timm==0.6.12
lightly==1.4.23
git+https://gitclone.com/github.com/lucasb-eyer/pydensecrf.git